Public v1.1 release
- **Added a dynamic delay mechanic**
Zefoy.com has a random delay for viewbot cycles between 2 and 3 minutes. For this reason, previously the bot had a static delay of 3 minutes to avoid breaking the code by submitting before the delay is over. Instead, a dynamic delay mechanic has been added, which reads the cooldown sent by Zefoy.com and applies that to it's own code. Because of this, the delay margin has been decreased from potentially 60 seconds to under 5 seconds.

- **Added a dynamic follower count system**
Previously, the bot would start at 0 views and add a static 1000 views by each view cycle. However, this wasn't accurate, so we implemented a feature that would actually read the view count of your video, do the math and display both current view count and the new view count (how many view the bot added through the last cycle)

- **Changed the way 'incorrect input' is managed**
Instead of sticking with the old 'press any key to exit' method upon entering an invalid input in any of the fields, the bot now returns you to the previous stage where you can retry and re-input your choice / link.